Renaming the junit package
